requirements.txt
multipledispatch==0.6.0
numpy==1.18.4
opencv-python==4.2.0.34
PyOpenGL==3.1.5
PyOpenGL-accelerate==3.1.5
pyrr==0.10.3
PySide2==5.14.2.1
shiboken2==5.14.2.1
six==1.15.0
